Reading Nook: Baked Lighting - VR 12/15/23
This was an assignment in my Materials and Lighting class. The assignment was to find an interior photo to mimic in VR. We weren't to focus on modeling, but find the assets we needed to create a like-image of the one we were working off of. We lit the scene and tried to match the general feel of the photo. We greyboxed the scene, found assets, dressed and lit the scene. 

Youtube Link: https://youtu.be/08ATl_aGPDg

Completed 12/15/2023