Forrest Dynamic Lighting - Scene Space - 11/01/23
This assignment was focused purely on lighting and feel as well, we were not expected to model, but to learn about lighting techniques. We were to find an outdoor image (could be a painting or photo) and recreate in Unreal. We found assets, dressed and lit the scene. 

Youtube Link: https://youtu.be/XiNCKInpLVs

(When I opened my project, a few things had broken, so it looks slightly different then the video from a year ago) 

Completed 11/01/2023