Dungeon Quest Adventure - 10/10/23
This was a lab that required us to make a Interface. This was the first time I made an interface or interactable blueprints, I used "E" as the interaction Key. 
I hadn't opened the Unreal file in a long time, when I did it was broken and the last room was gone. I will still include the Unreal file though. 

Youtube Link: https://youtu.be/IoyXwmHEe_4

Completed 10/10/2023