Roots of Knowledge VR Tour - 5/1/24
This was a service project. I made a tour of the Roots of Knowledge. The ultimate goal was to have interactions were you could listen to description of each stain glass. I didn't get that far but experimented with coding materials for one of the first times. I realized after I completed it that there was an easier way I could have done it. I think I should have made one stain glass and made children of the material and changed the picture to save time. Whoops! 

Youtube Link: https://youtu.be/Zm-sSh5nZI4